Fix: test result format changed, added ability to skip config copy in autotesting_init.py

+++ Projects/Autotesting/Scripts/autotesting_init.py
@@ -1,4 +1,4 @@
-#!/usr/bin/env python2.6
+#!/usr/bin/python
 #
 #  autotesting_init.py
 #  DAVA SDK
@@ -15,30 +15,38 @@ import platform;
 import shutil;
 import subprocess;
 
+arguments = sys.argv[1:]
+
+index_OS = 0
+index_Project = 1
+
+if 0 == len(arguments):
+	print 'Usage: ./autotesting_init.py [PlatformName] [ProjectName]'
+	exit(1)
+
 print "*** DAVA Initializing autotesting"
 
 print "platform.system: " + platform.system()
-print sys.argv
 
-index_OS = 1;
-index_Project = 2;
+
 
 def ignored_svn_files(adir,filenames):
     return [filename for filename in filenames if filename.endswith(".svn")]
 
 currentDir = os.getcwd(); 
 frameworkDir =  os.path.realpath(currentDir + "/../../../")
-projectDir = os.path.realpath(currentDir + "/../../../../" + sys.argv[index_Project])
+projectDir = os.path.realpath(currentDir + "/../../../../" + arguments[index_Project])
 print "Framework directory:" + frameworkDir
 print "Project directory:" + projectDir
 
-autotestingConfigSrcPath = os.path.realpath(currentDir + "/../Data/Config.h")
-autotestingConfigDestPath = os.path.realpath(frameworkDir + "/Sources/Internal/Autotesting/Config.h")
-if os.path.exists(autotestingConfigDestPath):    
-    print "delete " + autotestingConfigDestPath
-    os.remove(autotestingConfigDestPath)
-print "copy " + autotestingConfigSrcPath + " to " + autotestingConfigDestPath
-shutil.copy(autotestingConfigSrcPath, autotestingConfigDestPath)
+if 2 == len(arguments):
+    autotestingConfigSrcPath = os.path.realpath(currentDir + "/../Data/Config.h")
+    autotestingConfigDestPath = os.path.realpath(frameworkDir + "/Sources/Internal/Autotesting/Config.h")
+    if os.path.exists(autotestingConfigDestPath):    
+        print "delete " + autotestingConfigDestPath
+        os.remove(autotestingConfigDestPath)
+    print "copy " + autotestingConfigSrcPath + " to " + autotestingConfigDestPath
+    shutil.copy(autotestingConfigSrcPath, autotestingConfigDestPath)
 
 autotestingSrcFolder = os.path.realpath(projectDir + "/Autotesting")
 autotestingDestFolder = os.path.realpath(projectDir + "/Data/Autotesting")
@@ -46,7 +54,7 @@ autotestingDestFolder = os.path.realpath(projectDir + "/Data/Autotesting")
 scripts = ["/generate_id.py"]
 
 if (platform.system() == "Darwin"):
-    if (sys.argv[index_OS] == "iOS"):
+    if (arguments[index_OS] == "iOS"):
         scripts.append("/runOnDevice.sh")
         scripts.append("/floatsign.sh")
         scripts.append("/packipa.sh")
@@ -83,7 +91,7 @@ shutil.copytree(autotestingTestsSrcFolder, autotestingTestsDestFolder, ignore=ig
 
 os.chdir(autotestingSrcFolder)
 
-params = ["python", "./generate_id.py", sys.argv[index_Project], autotestingDestFolder]
+params = ["python", "./generate_id.py", arguments[index_Project], autotestingDestFolder]
 print "subprocess.call " + "[%s]" % ", ".join(map(str, params))
 subprocess.call(params)
